- 提取重复代码到单独方法中
- 使用增强 for 循环替换部分代码- 移除不必要的类型转换- 统一使用 Integer 类型

+++ src/client/inventory/MapleInventory.java
@@ -115,7 +115,7 @@ public class MapleInventory implements Iterable<IItem>, Serializable
         if (item.getPosition() < 0 && !this.type.equals((Object)MapleInventoryType.EQUIPPED)) {
             return;
         }
-        this.inventory.put(Short.valueOf(item.getPosition()), item);
+        this.inventory.put(item.getPosition(), item);
     }
     
     public void move(final short sSlot, final short dSlot, final short slotMax) {
